Change the type of the bitpacker window to size_t. This solves a problem for TI C64x+ DSPs, where sizeof(long)==8, but longs  really only contain 40 bits of data. It should still yield the preferred type on every other platform (e.g., 64  bits for x86-64), and as a bonus is now also correct for 64-bit Windows.
svn path=/trunk/theora/; revision=16715

+++ lib/bitpack.h
@@ -16,11 +16,12 @@
  ********************************************************************/
 #if !defined(_bitpack_H)
 # define _bitpack_H (1)
+# include <stddef.h>
 # include <limits.h>
 
 
 
-typedef unsigned long      oc_pb_window;
+typedef size_t             oc_pb_window;
 typedef struct oc_pack_buf oc_pack_buf;
